Move useExpenseStore to hooks folder and delete store folder
