+++ .gitignore
@@ -1,0 +1,6 @@
+node_modules/
+.expo/
+.agents/
+.claude/
+skills-lock.json
+*.log
