+++ App.tsx
@@ -1,7 +1,7 @@
 import "./global.css";
 import AppNavigator from './src/navigation/AppNavigator';
 import { useEffect } from 'react';
-import { useExpenseStore } from './src/features/expense/store/useExpenseStore';
+import { useExpenseStore } from './src/features/expense/hooks/useExpenseStore';
 
 export default function App() {
   const hydrate = useExpenseStore((state) => state.hydrate);
